Working with sdl_gamepad
