+++ lib/src/models/rover/controller.dart
@@ -92,7 +92,6 @@ class Controller extends Model {
   Future<void> _update([_]) async {
     final state = gamepad.getState();
     if (state == null) return;
-    // print(state.normalLeftX);
     controls.updateState(state);
     final messages = controls.parseInputs(state);
     for (final message in messages) {
