+++ lib/src/services/gamepad/gamepad.dart
@@ -2,8 +2,6 @@ import "dart:io";
 
 import "mock.dart";
 import "state.dart";
-import "windows.dart";
-import "sdl2.dart";
 import "sdl.dart";
 
 import "../service.dart";
@@ -24,8 +22,8 @@ abstract class Gamepad extends Service {
 
   /// Returns a functional instance of this class, or a mock on unsupported platforms.
   factory Gamepad.forPlatform(int index) => Platform.isWindows
-    ? Win32Gamepad(index)
-    : DashboardSdlGamepad(index);
+    ? DesktopGamepad(index)
+    : MockGamepad(index);
 
   /// Gets the current state of the gamepad, or null if it's not connected.
   GamepadState? getState();
