+++ pubspec.yaml
@@ -26,9 +26,7 @@ dependencies:
   collection: ^1.18.0
   sdl3: ^0.1.31
   ffi: ^2.1.3
-  sdl_gamepad:
-    # git: https://github.com/Levi-Lesches/sdl_gamepad
-    path: "D:\\flutter\\packages\\sdl_gamepad"
+  sdl_gamepad: ^1.0.1
 
 # Prefer to use `flutter pub add --dev packageName` rather than modify this section by hand.
 dev_dependencies:
